Move files to new screens folder structure
