+++ TodoListAdvanced/src/theme/colors.js
@@ -1,0 +1,2 @@
+
+export const primary = '#2288ee'
